listo, asistencia, registro  a clases y y optimizaciones

+++ rutas/diciplinas/deportista.js
@@ -1,10 +1,10 @@
-var expres = require('express')
-var autenticacion = require('../middelware/autenticacion')
+var expres = require('express');
+var autenticacion = require('../../middelware/autenticacion');
 var app = expres();
-var Deportista = require('../modelos/deportista');
-var Persona = require('../modelos/Personas');
-var contatoEmergencia = require('../modelos/contacEmergen');
-const { json } = require('express/lib/response');
+var Deportista = require('../../modelos/deportista');
+var Persona = require('../../modelos/Personas');
+var contatoEmergencia = require('../../modelos/contacEmergen');
+var Registro = require('../../modelos/clases/Registro');
 
 
 // ==============================
@@ -13,29 +13,28 @@ const { json } = require('express/lib/response');
 
 app.get('/', (req, res, next) => {
 
-
-
     var desde = req.query.desde || 0;
     desde = Number(desde);
-
-    Persona.find({},)
+    Deportista.find({},)
         .skip(desde)
-        .populate('Usuario', 'Nombre email')
-        .populate({
-            path: 'idperfil',
-            populate: [
-                {
-                    path: 'IdContacto'
-                },
-                {
-                    path: 'IdContacto2'
-                },
-                {
-                    path: 'Diciplina'
-                },
-            ]
-        }
-        )
+        .populate([
+            {
+                path: '_id',
+                select: '',
+                populate: [{
+                    path: 'Usuario',
+                    select: 'Nombre email'
+                }]
+            },
+            {
+                path: 'IdContacto',
+                select: 'Nombres Telefono',
+            },
+            {
+                path: 'IdContacto2',
+                select: 'Nombres Telefono',
+            }
+        ])
         .limit(5)
         .exec(
             (err, persona) => {
@@ -46,16 +45,16 @@ app.get('/', (req, res, next) => {
                         erros: err
                     });
                 }
-                // deportista.count({}, (err, conteo) => {
+                Deportista.count({}, (err, conteo) => {
 
+                    res.status(200).json({
+                        ok: true,
+                        personas: persona,
+                        total: conteo
+                    });
 
+                })
 
-                // })
-                res.status(200).json({
-                    ok: true,
-                    personas: persona,
-                    // total: conteo
-                });
 
             });
 
@@ -73,7 +72,6 @@ app.put('/:id', autenticacion.verificatoken, (req, res) => {
     var id = req.params.id;
     var body = req.body;
     var id_usuario = req.usuario._id;
-    var data;
     var deportistaDato = new Deportista({
         FechaNacimiento: body.FNacimiento,
         DepartamentNacimiento: body.departamentN,
@@ -107,60 +105,50 @@ app.put('/:id', autenticacion.verificatoken, (req, res) => {
 
     });
 
-    Persona.findByIdAndUpdate(id, persona)
-        .exec((err, deportista) => {
 
-
-            if (err) {
-                return res.status(500).json({
-                    ok: false,
-                    mensaje: 'Error al buscar deportista',
-                    erros: err
-                });
-            }
-
-            if (deportista = null || !deportista) {
-                return res.status(500).json({
-                    ok: false,
-                    mensaje: 'el deportista no esta en almacenado',
-                    erros: err
-                });
-            }
-            data = deportista;
-            Deportista.findByIdAndUpdate(id, deportistaDato)
-                .exec((err, deportista) => {
-
-
-                    if (err) {
-                        return res.status(500).json({
-                            ok: false,
-                            mensaje: 'Error al buscar deportista',
-                            erros: err
-                        });
-                    }
-
-                    if (deportista = null || !deportista) {
-                        return res.status(500).json({
-                            ok: false,
-                            mensaje: 'el deportista no esta en almacenado',
-                            erros: err
-                        });
-                    }
-                    data   = `${data}  ${deportista}`;
-                    res.status(500).json({
-                        ok: true,
-                        mensaje:  'deportista actualizado'
-
-                    });
-
-                });
-
-        });
-
-
-
-
+    Promise.all([
+        ActualizarPrsona(id, persona),
+        Actualizaralumno(id, deportistaDato),
+    ]).then(respuestas => {
+        res.status(200).json({
+            ok: true,
+            message: ' el alumno fue Actualizado Correcamente',
+            persona: respuestas[0],
+            Alumno: respuestas[1],
+        })
+    }
+    );
 });
+// ==============================
+// actualizar  los Contactos
+// ==============================
+app.put('/contacto/:id', autenticacion.verificatoken, (req, res) => {
+
+    var id = req.params.id;
+    var body = req.body;
+    var contacto = new contatoEmergencia({
+        _id: body.IdContacto,
+        Nombres: body.NombresContacto,
+        Apellidos: body.ApellidosContacto,
+        Barrio: body.BarrioContacto,
+        Direccion: body.direccionContacto,
+        Telefono: body.telefonoContacto,
+        Ocupacion: body.ocupacioncontacto,
+        email: body.emailcontacto
+    });
+
+
+
+    Actualizarcontacto(id, contacto).then(respuestas => {
+        res.status(200).json({
+            ok: true,
+            message: ' el contacto fue Actualizado Correcamente',
+            persona: respuestas,
+        })
+    }
+    );
+});
+
 // ==============================
 // ingresar deportista nuevo 
 // ==============================
@@ -182,7 +170,6 @@ app.post('/', autenticacion.verificatoken,
             Ocupacion: body.ocupacion,
             EPS: body.eps,
             email: body.email,
-            idperfil: body.id,
             Usuario: id_usuario
 
         });
@@ -192,7 +179,6 @@ app.post('/', autenticacion.verificatoken,
             DepartamentNacimiento: body.departamentN,
             MunicipioNacimiento: body.MunicipioN,
             Edad: body.edad,
-            Diciplina: body.Diciplina,
             IdContacto: body.IdContacto,
             IdContacto2: body.IdContacto2,
             padecimientos: body.padecimientos,
@@ -316,26 +302,25 @@ app.delete('/:id', autenticacion.verificatoken, (req, res) => {
 
     var id = req.params.id;
     var data;
-
-    Persona.findByIdAndRemove(id, (err, usuarioborrado) => {
+    Registro.findOne({ id_deportista: id }, (err, claseborrado) => {
 
 
         if (err) {
             return res.status(500).json({
                 ok: false,
-                mensaje: 'Error borar usuario',
+                mensaje: 'Error borar clase',
                 erros: err
             });
         }
-        if (!usuarioborrado) {
-            return res.status(400).json({
+        console.log(claseborrado);
+        if (claseborrado != null || claseborrado == {}) {
+            return res.status(205).json({
                 ok: false,
-                mensaje: 'persona con ' + id + ' no existe',
-                erros: { message: 'no existe el persona con ese id ' }
+                mensaje: 'el deportista con id: ' + id + ' tiene registrado alumnos  ',
+                erros: { message: 'borre los registros antes de borrar una clase ' }
             });
         }
-        data = usuarioborrado;
-        Deportista.findByIdAndRemove(id, (err, usuarioborrado) => {
+        Persona.findByIdAndRemove(id, (err, usuarioborrado) => {
 
 
             if (err) {
@@ -348,26 +333,137 @@ app.delete('/:id', autenticacion.verificatoken, (req, res) => {
             if (!usuarioborrado) {
                 return res.status(400).json({
                     ok: false,
-                    mensaje: 'deportes con ' + id + ' no existe',
-                    erros: { message: 'no existe el deportes con ese id ' }
+                    mensaje: 'persona con ' + id + ' no existe',
+                    erros: { message: 'no existe el persona con ese id ' }
                 });
             }
-            data = `${data} ${usuarioborrado}`;
-            res.status(200).json({
-                ok: true,
-                usuario: data
+            data = usuarioborrado;
+            Deportista.findByIdAndRemove(id, (err, usuarioborrado) => {
+
+
+                if (err) {
+                    return res.status(500).json({
+                        ok: false,
+                        mensaje: 'Error borar usuario',
+                        erros: err
+                    });
+                }
+                if (!usuarioborrado) {
+                    return res.status(400).json({
+                        ok: false,
+                        mensaje: 'deportes con ' + id + ' no existe',
+                        erros: { message: 'no existe el deportes con ese id ' }
+                    });
+                }
+                data = `${data} ${usuarioborrado}`;
+
             });
 
 
-
-
+        });
+        res.status(201).json({
+            ok: true,
+            mensaje: 'el  deportista fue borrado',
+            data: data
         });
 
+    });
+
+});
+
+// funciones de Actualizar datos////
+function ActualizarPrsona(id, dato) {
+    return new Promise((resolve, reject) => {
+        Persona.findByIdAndUpdate(id, dato)
+            // .populate('usuario', 'Nombre email')
+            .exec((err, persona) => {
+
+                if (err) {
+                    reject('error al cargar persona')
+                } else {
+                    resolve(persona)
+                }
+
+            });
+
+    });
+}
+function Actualizaralumno(id, dato) {
+    return new Promise((resolve, reject) => {
+        Deportista.findByIdAndUpdate(id, dato)
+            .exec((err, alumno) => {
+                if (err) {
+                    reject('error al cargar alumno')
+                } else {
+                    resolve(alumno)
+                }
+            });
+
+    });
+}
+function Actualizarcontacto(id, dato) {
+    return new Promise((resolve, reject) => {
+        contatoEmergencia.findByIdAndUpdate(id, dato)
+            .exec((err, contacto) => {
+                if (err) {
+                    reject('error al cargar alumno')
+                } else {
+                    resolve(contacto)
+                }
+            });
+
+    });
+}
+// funciones de guardado datos ////
+function GuardarAlumno(datos) {
+    return new Promise((resolve, reject) => {
+        datos.save((err, alumno) => {
+            if (err) {
+                reject('error al cargar alumno')
+            } else {
+                resolve(alumno)
+            }
+
+        });
+    });
+
+
+}
+function Guardarcontacto(datos) {
+    return new Promise((resolve, reject) => {
+        datos.save((err, contacto) => {
+            if (err) {
+                reject('error al cargar persona')
+            } else {
+                resolve(contacto)
+            }
+
+        });
+    });
+
+
+}
+// buscar contacto
+function buscarcontacto(busqueda) {
+    return new Promise((resolve, reject) => {
+        contatoEmergencia.find({ _id: busqueda })
+            // .populate('usuario', 'Nombre email')
+            .exec((err, contacto) => {
+
+                if (err) {
+                    reject('error al cargar contacto')
+                } else {
+                    resolve(contacto)
+                }
+
+            });
 
     });
 
 
-});
+}
+
+
